Product change

+++ ERPAnimalia/Models/DetailGrid.cs
@@ -10,14 +10,14 @@ namespace ERPAnimalia.Models
      public Guid IdProduct { get; set; }
      public string Codigo { get; set; }
      public string Descripcion1 { get; set; }
-     public double PrecioVenta { get; set; }
-     public double PrecioCosto { get; set; }
+     public decimal PrecioVenta { get; set; }
+     public decimal PrecioCosto { get; set; }
         public decimal Cantidad { get; set; }
      public decimal kg { get; set; }
-     public double Descuento { get; set; }
+     public decimal Descuento { get; set; }
      public string Subtotal { get; set; }
      public decimal Total { get; set; }
-     public double Porcentage { get; set; }
+     public decimal Porcentage { get; set; }
      public int CategoryItem { get; set; }
      public int SubCategoryItem { get; set; }
     }
